requirements.txt
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.12.0
langchain==0.0.184
pyyaml==6.0.2
tqdm>=4.62.3
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
ollama>=0.1.0
interruptingcow==0.8
transformers>=4.38.0
torch>=2.2.0
flask>=2.3.0
werkzeug>=2.3.0
